secure course + location add/edit/delete views

+++ courses/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.views.generic import ListView, DetailView
 from .models import Course
 from django.db.models import Q
@@ -62,9 +63,12 @@ class BeginnerListView(ListView):
     def get_queryset(self):
         return Course.objects.all().filter(level='Beginner')
 
-
+@login_required
 def add_course(request):
 # adds course to the site
+    if not request.user.is_superuser:
+        messages.error(request, 'Tha sinn duilich. This page is only for site admin.')
+        return redirect(reverse('home'))
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES)
         if form.is_valid():
@@ -83,9 +87,12 @@ def add_course(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_course(request, course_id):
 # edits a course on the site
+    if not request.user.is_superuser:
+        messages.error(request, 'Tha sinn duilich. This page is only for site admin.')
+        return redirect(reverse('home'))
     course = get_object_or_404(Course, pk=course_id)
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES, instance=course)
@@ -108,8 +115,12 @@ def edit_course(request, course_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_course(request, course_id):
     # delete course store site
+    if not request.user.is_superuser:
+        messages.error(request, 'Tha sinn duilich. This page is only for site admin.')
+        return redirect(reverse('home'))
     course = get_object_or_404(Course, pk=course_id)
     course.delete()
     messages.success(request, 'Sgonneil! Course deleted!')
