+++ locations/views.py
@@ -1,8 +1,10 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.views.generic import ListView, DetailView
 from .models import Location
 from .forms import LocationForm
+
 
 # Create your views here.
 class LocationListView(ListView):
@@ -17,8 +19,12 @@ class LocationDetailView(DetailView):
     context_object_name = 'location'
     template_name = 'locations/location_detail.html'
 
+@login_required
 def add_location(request):
 # adds location to the site
+    if not request.user.is_superuser:
+        messages.error(request, 'Tha sinn duilich. This page is only for site admin.')
+        return redirect(reverse('home'))
     if request.method == 'POST':
         form = LocationForm(request.POST, request.FILES)
         if form.is_valid():
@@ -38,8 +44,12 @@ def add_location(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_location(request, location_id):
 # edits a location on the site
+    if not request.user.is_superuser:
+        messages.error(request, 'Tha sinn duilich. This page is only for site admin.')
+        return redirect(reverse('home'))
     location = get_object_or_404(Location, pk=location_id)
     if request.method == 'POST':
         form = LocationForm(request.POST, request.FILES, instance=location)
@@ -61,8 +71,13 @@ def edit_location(request, location_id):
 
     return render(request, template, context)
 
+
+@login_required
 def delete_location(request, location_id):
     # delete location store site
+    if not request.user.is_superuser:
+        messages.error(request, 'Tha sinn duilich. This page is only for site admin.')
+        return redirect(reverse('home'))
     location = get_object_or_404(Location, pk=location_id)
     location.delete()
     messages.success(request, 'Sgonneil! Location deleted!')
